move language related extensions to :core:l10n
